django/requirements.txt
django==2.2.5
gunicorn==19.9.0
psycopg2==2.8.4
dj-database-url==0.5.0
django-cors-headers==3.1.1
pytest==5.2.2
pytest-cov==2.8.1
pytest-django==3.6.0
selenium==3.141.0
djangorestframework==3.10.3
Pillow==6.2.1
tensorflow==1.15.0
keras==2.2.4
scikit-learn==0.21.3